requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.1
pillow==11.0.0
pytube==15.0.0
pytubefix==8.2.0
typing_extensions==4.12.2
